update [component]: Checkbox, FiatValue, +ss cl wl
